move code to subdirectory, use "fake" names
